requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
flask==3.1.2
idna==3.11
importlib-metadata==8.7.1
itsdangerous==2.2.0
jinja2==3.1.6
mailjet-rest==1.4.0
markupsafe==3.0.3
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
werkzeug==3.1.5
zipp==3.23.0
gunicorn